2 day meal history + basal lock Customizations - Update build_loop.yml
Adding 2 new customizations: 

Display Two Days of Meal History
This customization extends the Carbohydrates history page so the earliest data shown is from midnight a day ago. This is a shorter period that you get when you choose Display a Week of Meal History, but it loads faster after quitting and restarting the app. Only one of these can be chosen.

+ Basal Lock